rename and refactor Workflow package to Hybrid package
