+++ fuel/packages/hybrid/bootstrap.php
@@ -1,0 +1,8 @@
+<?php
+
+Autoloader::add_classes(array (
+	'Hybrid\\Request'		=> __DIR__.'/classes/request.php',
+	'Hybrid\\Input'			=> __DIR__.'/classes/input.php'
+));
+
+/* End of file bootstrap.php */
