--- fuel/packages/workflow/bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-Autoloader::add_classes(array (
-	'Workflow\\Request'		=> __DIR__.'/classes/request.php',
-	'Workflow\\Input'		=> __DIR__.'/classes/input.php'
-));
-
-/* End of file bootstrap.php */
